저장소 생성 컴포넌트 분리
